pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-support-agent"
version = "0.2.0"
description = "LLM-powered support automation agent backend (Ollama)"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "structlog>=24.1",
  "prometheus_client>=0.20",
  "celery>=5.4",
  "redis>=5.0",
  "httpx>=0.27",           # Ollama HTTP API
  "orjson>=3.10",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
